# Speed up skipped adapter specs