Minor tweak to speed up skipping of adapter specs in local dev if underlying sever is not available. Currently each example has to wait for the error, which could be ~1 second per example. This sets a global and just short-circuits the skip if one of the pervious examples in the example group fail.

Before:
```
Finished in 50.18 seconds (files took 5.08 seconds to load)
1719 examples, 0 failures, 69 pending
```

After:
```
Finished in 10.86 seconds (files took 2.63 seconds to load)
1719 examples, 0 failures, 69 pending
```